Refactor fetching code to Drawer component
