+++ client/src/components/OptionsDrawer/OptionsDrawer.tsx
@@ -1,23 +1,34 @@
+import React, { useState, useEffect } from 'react';
 import Typography from '@material-ui/core/Typography';
 import ListItem from '@material-ui/core/ListItem';
 import Divider from '@material-ui/core/Divider';
 import List from '@material-ui/core/List';
 import Drawer from '@material-ui/core/Drawer';
-import React from 'react';
 import ListItemText from '@material-ui/core/ListItemText';
 import ListItemIcon from '@material-ui/core/ListItemIcon';
 import SwapHorizontalCircleOutlinedIcon from '@material-ui/icons/SwapHorizontalCircleOutlined';
 import useStyles from './optionsDrawerStyles';
 import { IBoard } from '../../interface/Board';
+import getUserBoards from '../../helpers/APICalls/getUserBoards';
 
 interface Props {
   open: boolean;
-  boards: Array<IBoard>;
+  // boards: Array<IBoard>;
   setOpen: React.Dispatch<React.SetStateAction<boolean>>;
 }
 
-const OptionsDrawer = ({ open, boards, setOpen }: Props): JSX.Element => {
+const OptionsDrawer = ({ open, setOpen }: Props): JSX.Element => {
   const classes = useStyles();
+  const [boards, setBoards] = useState<Array<IBoard>>([]);
+
+  const getAllUserBoards = async () => {
+    const { boards } = await getUserBoards();
+    if (boards) setBoards(boards);
+  };
+
+  useEffect(() => {
+    getAllUserBoards();
+  }, []);
 
   return (
     <Drawer anchor={'right'} open={open} onClose={setOpen}>
